backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
django-cors-headers==4.4.0

